[Chore][Core/Dashboard] Split the implementation of ServeHead and ServeAgent
